Fix build failure on Podman vendor
Basically Windows build is failing when hitting unix contants in
cgroups.

[NO NEW TESTS ADDED] Just shifting code around.

Added build-cross to cirrus tests, which looks like it was not being
run.
