+++ pkg/cgroups/cgroups_unsupported.go
@@ -2,6 +2,10 @@
 
 package cgroups
 
+import (
+	"os"
+)
+
 // IsCgroup2UnifiedMode returns whether we are running in cgroup 2 cgroup2 mode.
 func IsCgroup2UnifiedMode() (bool, error) {
 	return false, nil
@@ -11,4 +15,8 @@ func IsCgroup2UnifiedMode() (bool, error) {
 // current cgroup.
 func UserOwnsCurrentSystemdCgroup() (bool, error) {
 	return false, nil
+}
+
+func rmDirRecursively(path string) error {
+	return os.RemoveAll(path)
 }
